Add optional define for having the capsuile pointer on the heap.
This is for multiple c source file c extensions where the code that uses the capsule pointer is in a separate source file than the module init function, and the module init function being the only place where the capsule init is ran inside of.

Originally I was going to have a define check that the user can define their own variable for all the defines to use below, but thought against it as the change is bigger than simply doing this instead. Also with it being an optional define it should be fine.