Use vendoring as the distribution method (#59)
This eliminates function tables entirely, and instead ships the entire
PyAwaitable source code as a header file. This may seem bad at first,
but has a lot of nice benefits for us:

- PyAwaitable is no longer needed as a runtime dependency, as it will be
completely baked into an extension.
- In turn, there's no chance of runtime collisions between installed
versions of PyAwaitable.
- Thus, it's completely safe to pin a PyAwaitable version to an exact
number without worrying about compatibility.

On the other hand, there a few downsides:

- Our `ctypes` bindings need to be written in favor of a C extension.
- There is some added complexity with how things are initialized,
because a virtual extension has to be created and stored on the
interpreter state.
- We're now completely reliant on a vendoring script, which could prove
to be maintenance heavy. (But I doubt it will be worse than capsules.)
- Debugging will be screwed up a little bit downstream, because function
names are mangled.

Closes #30 and #47.

---------

Co-authored-by: AraHaan <seandhunt_7@yahoo.com>