Tests: Add new tests for the value APIs (#66)
With this, I also decided to remove the integer value functions. They're
a bit buggy because they only really work with longs, which doesn't
cover enough use cases. I could add integer APIs for every single type
of integer, but it's really not worth it compared to just allocating an
integer and storing it via the arbitrary values.