Core: Add `PyAwaitable_AddExpr` for convenience (#81)
Inspired by `PyModule_Add`. This function is equivalent to
`PyAwaitable_AddAwait`, but returns `NULL` without setting an exception
when given a `NULL` coroutine, and steals a reference to the coroutine
when it is non-`NULL`. This makes it possible to directly use C API
functions in the single `PyAwaitable_AddExpr` call.